refactor(value): use lifecycle.onMounted helper
